GH-1841 - Fix tests broken by inheritance refactor
