+++ allensdk/test/brain_observatory/behavior/test_trials_processing.py
@@ -22,8 +22,9 @@ def test_get_ori_info_from_trial(behavior_experiment_id, ti, expected, exception
     -----
     - i may be rewriting code here but its more a sanity check really...
     """
+    lims_api = BehaviorLimsApi(behavior_session_id=behavior_experiment_id)
     stim_output = pd.read_pickle(
-        BehaviorLimsApi(behavior_session_id=behavior_experiment_id).get_behavior_stimulus_file()
+        lims_api.raw_data_api.get_behavior_stimulus_file()
     )
     trial_log = stim_output['items']['behavior']['trial_log']
